+++ ScriptReloader/App.xaml.cs
@@ -1,3 +1,4 @@
+using System.Reflection;
 using System.Windows;
 using Microsoft.Extensions.Configuration;
 using ScriptReloader.Models;
@@ -6,11 +7,23 @@ namespace ScriptReloader;
 
 public partial class App : Application
 {
+    private const string EmbeddedAppSettingsName = "ScriptReloader.appsettings.json";
+
     private void OnStartup(object sender, StartupEventArgs e)
     {
         var builder = new ConfigurationBuilder()
-            .SetBasePath(AppContext.BaseDirectory)
-            .AddJsonFile("appsettings.json", optional: true, reloadOnChange: false)
+            .SetBasePath(AppContext.BaseDirectory);
+
+        var asm = Assembly.GetExecutingAssembly();
+        using (var stream = asm.GetManifestResourceStream(EmbeddedAppSettingsName))
+        {
+            if (stream is not null)
+            {
+                builder.AddJsonStream(stream);
+            }
+        }
+
+        builder.AddJsonFile("appsettings.json", optional: true, reloadOnChange: false)
             .AddUserSecrets(typeof(App).Assembly, optional: true)
             .AddEnvironmentVariables();